[feat]기본 CR 구현

+++ src/main/kotlin/com/example/onlineshoping/project/domain/dto/request/CreateBuyerRequest.kt
@@ -5,5 +5,5 @@ data class CreateBuyerRequest(
     val name:String,
     val phoneNumber:String,
     val addressType: String,
-    val buyerBalance:Int
+    val buyerBalance:Long
 )
